removed profile_reducer and added categories into sets reducer and added user info into user_reducer
