+++ project_setup/mern_setup/client/src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
 import userReducer from './user_reducer';
-import profileReducer from './profile_reducer';
-
 import setsReducer from './sets_reducer';
 import cardReducer from './card_reducer';
 import modalReducer from './modal_reducer';
@@ -11,7 +9,6 @@ import createCardReducer from './add_card_reducer';
 
 const rootReducer = combineReducers ({
     user: userReducer,
-    profile: profileReducer,
     form: formReducer,
     sets:setsReducer,
     card: cardReducer,
