+++ project_setup/mern_setup/client/src/reducers/sets_reducer.js
@@ -2,6 +2,7 @@ import types from '../actions/types';
 
  const DEFAULT_STATE = {
     category: "",
+    categories: [],
     sets: [] 
 };
  export default (state = DEFAULT_STATE, action) => {
@@ -12,8 +13,12 @@ import types from '../actions/types';
         const category = action.payload.data.sets.length
                         ? action.payload.data.sets[0].category
                         : "No data available";
-        return {...state,sets, category};
+            return {...state,sets, category};
+        case types.SORT_ALPHABETICAL:
+            return {...state, categories: action.payload };
+        case types.SORT_BY_LATEST:
+            return {...state, categories: action.payload };  
         default:
-        return state;
+            return state;
     }
 }
