+++ project_setup/mern_setup/server/config/db.js
@@ -1,13 +1,13 @@
-module.exports = {
-    'host': 'localhost',
-    'user': 'root',
-    'password': 'root',
-    'database': 'pocketcards'
-}
-
 // module.exports = {
-//     'host': 'tiffanyykim.com',
-//     'user': 'pocket_cards',
-//     'password': 'mRN7Us9^7p5YdCwu',
+//     'host': 'localhost',
+//     'user': 'root',
+//     'password': 'root',
 //     'database': 'pocketcards'
 // }
+
+module.exports = {
+    'host': 'tiffanyykim.com',
+    'user': 'pocket_cards',
+    'password': 'mRN7Us9^7p5YdCwu',
+    'database': 'pocketcards'
+}
